Refactor directories to match transformer design pattern
