+++ generators/ui-base.js
